requirements.txt
requests
mysql-connector-python
sshtunnel
python-dotenv
pytest
